effects: Refactor hex patterns to use native math, bounds fixes
Migrates patterns to hex/ directory, fixes bounds, optimizes performance for ESP32, and removes use of heap in hot paths.

+++ include/effects/hex/PatternHexFireworks.h
@@ -15,8 +15,8 @@ struct HexParticle {
     HexCoord position;
     HexCoord velocity;
     CRGB color;
-    uint8_t life;
-    uint8_t maxLife;
+    int life;
+    int maxLife;
 };
 
 class PatternHexFireworks : public EffectWithId<PatternHexFireworks>
